Tidy up code generation for paramter types
