+++ inst/cpp_generation/model_input.csv
@@ -1,38 +1,38 @@
-r_name,cpp_name,type,input_when,struct,value,convert_base,dims,dim1,dim2,dim3,dim4
-basepop,base_pop,real_type,,Demography,,FALSE,2,base.pAG,base.NS,,
-Sx,survival_probability,real_type,,Demography,,FALSE,3,base.pAG + 1,base.NS,proj_years,
-netmigr_adj,net_migration,real_type,,Demography,,FALSE,3,base.pAG,base.NS,proj_years,
-asfr,age_specific_fertility_rate,real_type,,Demography,,FALSE,2,options.p_fertility_age_groups,proj_years,,
-births_sex_prop,births_sex_prop,real_type,,Demography,,FALSE,2,base.NS,proj_years,,
-incidinput,total_rate,real_type,,Incidence,,FALSE,1,proj_years,,,
-incrr_age,relative_risk_age,real_type,,Incidence,,FALSE,3,base.pAG - options.p_idx_hiv_first_adult,base.NS,proj_years,
-incrr_sex,relative_risk_sex,real_type,,Incidence,,FALSE,1,proj_years,,,
-cd4_mort,cd4_mortality,real_type,,NaturalHistory,,FALSE,3,base.hDS,base.hAG,base.NS,
-cd4_prog,cd4_progression,real_type,,NaturalHistory,,FALSE,3,base.hDS - 1,base.hAG,base.NS,
-artcd4elig_idx,idx_hm_elig,int,,Art,,TRUE,1,proj_years + 1,,,
-cd4_initdist,cd4_initial_distribution,real_type,,NaturalHistory,,FALSE,3,base.hDS,base.hAG,base.NS,
-art_mort,mortality,real_type,,Art,,FALSE,4,base.hTS,base.hDS,base.hAG,base.NS
-artmx_timerr,mortaility_time_rate_ratio,real_type,,Art,,FALSE,2,base.hTS,proj_years,,
-art_dropout,dropout,real_type,,Art,,FALSE,1,proj_years,,,
-art15plus_num,adults_on_art,real_type,,Art,,FALSE,2,base.NS,proj_years,,
-art15plus_isperc,adults_on_art_is_percent,int,,Art,,FALSE,2,base.NS,proj_years,,
-,h_art_stage_dur,real_type,,Art,0.5,FALSE,1,base.hTS - 1,,,
-scale_cd4_mort,scale_cd4_mortality,int,,NaturalHistory,,FALSE,1,1,,,
-art_alloc_mxweight,initiation_mortality_weight,real_type,,Art,,FALSE,1,1,,,
-paed_incid_input,hc_nosocomial,real_type,ModelVariant::run_child_model,Children,,FALSE,1,proj_years,,,
-paed_cd4_dist,hc1_cd4_dist,real_type,ModelVariant::run_child_model,Children,,FALSE,1,children.hc2DS,,,
-paed_cd4_transition,hc_cd4_transition,real_type,ModelVariant::run_child_model,Children,,FALSE,2,children.hc2DS,children.hc1DS,,
-paed_cd4_mort,hc1_cd4_mort,real_type,ModelVariant::run_child_model,Children,,FALSE,3,children.hc1DS,children.hcTT,children.hc1AG,
-adol_cd4_mort,hc2_cd4_mort,real_type,ModelVariant::run_child_model,Children,,FALSE,3,children.hc2DS,children.hcTT,children.hc2AG,
-paed_cd4_prog,hc1_cd4_prog,real_type,ModelVariant::run_child_model,Children,,FALSE,1,children.hc1DS,,,
-adol_cd4_prog,hc2_cd4_prog,real_type,ModelVariant::run_child_model,Children,,FALSE,1,children.hc2DS,,,
-ctx_effect,ctx_effect,real_type,ModelVariant::run_child_model,Children,,FALSE,1,1,,,
-ctx_val,ctx_val,real_type,ModelVariant::run_child_model,Children,,FALSE,1,proj_years,,,
-paed_art_elig_age,hc_art_elig_age,real_type,ModelVariant::run_child_model,Children,,FALSE,1,proj_years,,,
-paed_art_elig_cd4,hc_art_elig_cd4,real_type,ModelVariant::run_child_model,Children,,TRUE,2,options.p_idx_hiv_first_adult,proj_years,,
-mort_art_rr,hc_art_mort_rr,real_type,ModelVariant::run_child_model,Children,,FALSE,3,base.hTS,options.p_idx_hiv_first_adult,proj_years,
-paed_art_mort,hc1_art_mort,real_type,ModelVariant::run_child_model,Children,,FALSE,3,children.hc1DS,base.hTS,children.hc1AG,
-adol_art_mort,hc2_art_mort,real_type,ModelVariant::run_child_model,Children,,FALSE,3,children.hc2DS,base.hTS,children.hc2AG,
-artpaeds_isperc,hc_art_isperc,int,ModelVariant::run_child_model,Children,,FALSE,1,proj_years,,,
-paed_art_val,hc_art_val,real_type,ModelVariant::run_child_model,Children,,FALSE,1,proj_years,,,
-init_art_dist,hc_art_init_dist,real_type,ModelVariant::run_child_model,Children,,FALSE,2,options.p_idx_hiv_first_adult,proj_years,,
+r_name,cpp_name,type,input_when,value,convert_base,dims,dim1,dim2,dim3,dim4
+basepop,Demography.base_pop,real_type,,,FALSE,2,base.pAG,base.NS,,
+Sx,Demography.survival_probability,real_type,,,FALSE,3,base.pAG + 1,base.NS,proj_years,
+netmigr_adj,Demography.net_migration,real_type,,,FALSE,3,base.pAG,base.NS,proj_years,
+asfr,Demography.age_specific_fertility_rate,real_type,,,FALSE,2,options.p_fertility_age_groups,proj_years,,
+births_sex_prop,Demography.births_sex_prop,real_type,,,FALSE,2,base.NS,proj_years,,
+incidinput,Incidence.total_rate,real_type,,,FALSE,1,proj_years,,,
+incrr_age,Incidence.relative_risk_age,real_type,,,FALSE,3,base.pAG - options.p_idx_hiv_first_adult,base.NS,proj_years,
+incrr_sex,Incidence.relative_risk_sex,real_type,,,FALSE,1,proj_years,,,
+cd4_mort,NaturalHistory.cd4_mortality,real_type,,,FALSE,3,base.hDS,base.hAG,base.NS,
+cd4_prog,NaturalHistory.cd4_progression,real_type,,,FALSE,3,base.hDS - 1,base.hAG,base.NS,
+artcd4elig_idx,Art.idx_hm_elig,int,,,TRUE,1,proj_years + 1,,,
+cd4_initdist,NaturalHistory.cd4_initial_distribution,real_type,,,FALSE,3,base.hDS,base.hAG,base.NS,
+art_mort,Art.mortality,real_type,,,FALSE,4,base.hTS,base.hDS,base.hAG,base.NS
+artmx_timerr,Art.mortaility_time_rate_ratio,real_type,,,FALSE,2,base.hTS,proj_years,,
+art_dropout,Art.dropout,real_type,,,FALSE,1,proj_years,,,
+art15plus_num,Art.adults_on_art,real_type,,,FALSE,2,base.NS,proj_years,,
+art15plus_isperc,Art.adults_on_art_is_percent,int,,,FALSE,2,base.NS,proj_years,,
+,Art.h_art_stage_dur,real_type,,0.5,FALSE,1,base.hTS - 1,,,
+scale_cd4_mort,NaturalHistory.scale_cd4_mortality,int,,,FALSE,1,1,,,
+art_alloc_mxweight,Art.initiation_mortality_weight,real_type,,,FALSE,1,1,,,
+paed_incid_input,Children.hc_nosocomial,real_type,ModelVariant::run_child_model,,FALSE,1,proj_years,,,
+paed_cd4_dist,Children.hc1_cd4_dist,real_type,ModelVariant::run_child_model,,FALSE,1,children.hc2DS,,,
+paed_cd4_transition,Children.hc_cd4_transition,real_type,ModelVariant::run_child_model,,FALSE,2,children.hc2DS,children.hc1DS,,
+paed_cd4_mort,Children.hc1_cd4_mort,real_type,ModelVariant::run_child_model,,FALSE,3,children.hc1DS,children.hcTT,children.hc1AG,
+adol_cd4_mort,Children.hc2_cd4_mort,real_type,ModelVariant::run_child_model,,FALSE,3,children.hc2DS,children.hcTT,children.hc2AG,
+paed_cd4_prog,Children.hc1_cd4_prog,real_type,ModelVariant::run_child_model,,FALSE,1,children.hc1DS,,,
+adol_cd4_prog,Children.hc2_cd4_prog,real_type,ModelVariant::run_child_model,,FALSE,1,children.hc2DS,,,
+ctx_effect,Children.ctx_effect,real_type,ModelVariant::run_child_model,,FALSE,1,1,,,
+ctx_val,Children.ctx_val,real_type,ModelVariant::run_child_model,,FALSE,1,proj_years,,,
+paed_art_elig_age,Children.hc_art_elig_age,real_type,ModelVariant::run_child_model,,FALSE,1,proj_years,,,
+paed_art_elig_cd4,Children.hc_art_elig_cd4,real_type,ModelVariant::run_child_model,,TRUE,2,options.p_idx_hiv_first_adult,proj_years,,
+mort_art_rr,Children.hc_art_mort_rr,real_type,ModelVariant::run_child_model,,FALSE,3,base.hTS,options.p_idx_hiv_first_adult,proj_years,
+paed_art_mort,Children.hc1_art_mort,real_type,ModelVariant::run_child_model,,FALSE,3,children.hc1DS,base.hTS,children.hc1AG,
+adol_art_mort,Children.hc2_art_mort,real_type,ModelVariant::run_child_model,,FALSE,3,children.hc2DS,base.hTS,children.hc2AG,
+artpaeds_isperc,Children.hc_art_isperc,int,ModelVariant::run_child_model,,FALSE,1,proj_years,,,
+paed_art_val,Children.hc_art_val,real_type,ModelVariant::run_child_model,,FALSE,1,proj_years,,,
+init_art_dist,Children.hc_art_init_dist,real_type,ModelVariant::run_child_model,,FALSE,2,options.p_idx_hiv_first_adult,proj_years,,
